# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NRWA API Documentation",
        "description": "REST API dokumentacija za NRWA projekt",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/v1/dispatchers": {
            "get": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Get list of dispatchers",
                "operationId": "d193b74eccd768ead409d2eda396c439",
                "responses": {
                    "200": {
                        "description": "List of dispatchers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dispatchers"
                ],
                "summary": "Create new dispatcher",
                "operationId": "6ddf4e52dd56f1dd6f4a5ea2ad147f60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ivan Horvat"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+38761234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispatcher created successfully"
                    }
                }
            }
        },
...
        "/api/v1/locations": {
            "get": {
                "tags": [
                    "Locations"
                ],
                "summary": "Get list of locations",
                "operationId": "42fc30e97773855c314df24aee584ce5",
                "responses": {
                    "200": {
                        "description": "List of locations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Locations"
                ],
                "summary": "Create new location",
                "operationId": "32ee553d585d8b4295d1721b072d79af",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Konzum Mostar"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Mostar, BiH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    }
                }
            }
        },
...
        "/api/v1/managers": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Get list of managers",
                "operationId": "a25ce70021d9694646a7a1e8e43c499d",
                "responses": {
                    "200": {
                        "description": "List of managers"
                    }
                }
            },
            "post": {
                "tags": [
                    "Managers"
                ],
                "summary": "Create new manager",
                "operationId": "d79084b9cca22f2c9c76518ecd00cc81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Marko Marković"
                                    },
                                    "location_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Manager created successfully"
                    }
                }
            }
        },
...
    },
...
}